utils/details_translator/requirements.txt
# Database
psycopg2-binary>=2.9.5

# Environment variables
python-dotenv>=0.21.0

# Image processing
Pillow>=9.3.0
pytesseract>=0.3.10
requests>=2.28.1
beautifulsoup4>=4.11.1

# OpenAI API
openai>=1.0.0